Rename interpreter module.
