+++ rust/src/nasl/interpreter/call.rs
@@ -6,7 +6,7 @@ use std::collections::HashMap;
 
 use super::interpreter::{InterpretResult, Interpreter};
 use crate::nasl::{
-    inter::{FunctionCallError, InterpretError, InterpretErrorKind},
+    interpreter::{FunctionCallError, InterpretError, InterpretErrorKind},
     syntax::{Statement, StatementKind},
     utils::lookup_keys::FC_ANON_ARGS,
     ContextType, NaslValue,
