+++ rust/src/nasl/interpreter/error.rs
