+++ rust/src/nasl/interpreter/forking_interpreter.rs
