+++ rust/src/nasl/interpreter/tests/local_var.rs
