+++ rust/src/nasl/test_utils.rs
@@ -20,7 +20,8 @@ use crate::{
 use futures::{Stream, StreamExt};
 
 use super::{
-    builtin::ContextFactory, inter::ForkingInterpreter, inter::InterpretErrorKind, utils::Executor,
+    builtin::ContextFactory, interpreter::ForkingInterpreter, interpreter::InterpretErrorKind,
+    utils::Executor,
 };
 
 // The following exists to trick the trait solver into
